# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Audio Guide 2026 API
  version: 1.13.0
  description: Canonical API for Audio Guide 2026. Includes fail-closed caching rules.
servers:
  - url: https://audiogid-api.vercel.app/v1
    description: Production Server
paths:
...
  /auth/me:
    get:
      summary: Get Current User
      operationId: me
      security:
        - BearerAuth: []
      tags: [Auth]
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /auth/logout:
    post:
      summary: Logout
      description: Revokes current access token and optional refresh token.
      operationId: logout
      security:
        - BearerAuth: []
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshReq' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string } } }
  /auth/login/sms/init:
    post:
      summary: Init SMS Login
      operationId: loginSmsInit
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhoneInit' }
      responses:
        '200':
          description: SMS sent
          content:
            application/json:
              schema: { type: object, properties: { status: { type: string }, detail: { type: string } } }
  /auth/login/sms/verify:
    post:
      summary: Verify SMS Login
      operationId: loginSmsVerify
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhoneVerify' }
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
  /auth/login/telegram:
    post:
      summary: Telegram Login
      operationId: loginTelegram
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelegramLogin' }
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
  /auth/refresh:
    post:
      summary: Refresh Access Token
      operationId: refreshToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshReq' }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Invalid token

  # --- Deletion ---
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    YookassaSignature:
      type: apiKey
      in: header
      name: X-Yookassa-Signature
  schemas:
...
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        refresh_token: { type: string, nullable: true }
        token_type: { type: string }
    RefreshReq:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
    PhoneInit:
      type: object
      required: [phone]
      properties:
        phone: { type: string }
    PhoneVerify:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string }
        code: { type: string }
    TelegramLogin:
      type: object
      required: [id, auth_date, hash]
      properties:
        id: { type: string }
        first_name: { type: string, nullable: true }
        username: { type: string, nullable: true }
        photo_url: { type: string, nullable: true }
        auth_date: { type: string }
        hash: { type: string }